CDP-1077: Add Notification When OpenNet is Detected
Added detect-browser package for determining browser name.
Added Alert component that displays a dismissable red bar with text prompting to switch to Chrome.
Added functionality supporting Alert that checks browser and reaches out to the API to detect OpenNet IPs.
Added Alert to Header.
Also connected App.js so that a class could be added to the containing element when the Alert is displayed so that the page body can be shifted down more.